\name{get_index}
\alias{get_index}
\title{Get the index associated with a marker in a raw parsed file}
\usage{
  get_index(raw.table, markers)
}
\arguments{
  \item{raw.table}{Lines of a raw table}

  \item{markers}{Regexes representing the start of a
  section (row) in a table}
}
\value{
  A vector of indices
}
\description{
  Private
}

