\name{parse_file}
\alias{parse_file}
\title{Parse a single file based on the file name}
\usage{
  parse_file(file.name)
}
\arguments{
  \item{file.name}{The situation report to parse}
}
\value{
  A data.frame representing a single situation report
}
\description{
  Parse a single file based on the file name
}

